Add displayLuxonDate to ui-functions
