+++ index.html
@@ -7,6 +7,8 @@
   <title>Awesome Books</title>
   <link rel="stylesheet" href="styles.css">
   <link rel="stylesheet" href="https://bootswatch.com/5/flatly/bootstrap.min.css">
+  <link rel="shortcut icon" href="./assets/book-icon.png" type="image/x-icon">
+  <script src="app.js" type="module" defer></script>
   <script src="https://cdn.jsdelivr.net/npm/luxon@1.27.0/build/global/luxon.min.js"></script>
 </head>
 <body>
@@ -68,6 +70,5 @@
   <footer class="myfoot bg-primary p-2 mt-5 ">
       <p class="m-2">copyright &copy; 2021, All Rights Reserved</p>
   </footer>
-  <script src="app.js" type="module"></script>
 </body>
 </html>
